REFACTOR(frontend): Phase 4 Frontend Polish
- Extract `useMailboxCallbacksValue` hook from `SpaApp.tsx`: the 35-line
  inline `mailboxCallbacksValue` construction is moved to
  `apps/web/src/hooks/useMailboxCallbacksValue.ts`, decoupling callback
  wiring from render logic and making `SpaApp` significantly shorter.
- Fix Title Case violation in `AnalyticsView`: subtitle was
  "Usage trends and processing stats"; corrected to
  "Usage Trends And Processing Stats" per CLAUDE.md convention.
- Memoize `selectedApplication` in `MailboxesView`: wrapped the
  `applications.find()` call in `useMemo` to avoid recomputing on
  every render.
